requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
django-redis==5.4.0
mozilla-django-oidc==3.0.0
django-allauth==0.61.1
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7 